requirements.txt
aws-cdk.aws_s3
aws-cdk.aws_s3_deployment
aws-cdk.aws_cloudfront
aws-cdk.aws_cloudfront_origins
aws-cdk.core